Extract pure audio codec helpers into audio-pcm-codec module
Pure code, no React, no DOM event listeners — directly unit-testable
in isolation. Moves out of useAudioDictationRecorder:

- CANONICAL_AUDIO_SAMPLE_RATE_HZ / WAV_HEADER_BYTES / BYTES_PER_SAMPLE
- AUDIO_BARS_COUNT
- AudioDictationDiagnostics type
- writeAscii (internal)
- createCanonicalWavBytesFromPcm
- resampleMonoFloat32ToPcm16
- mediaTrackSettingsToDiagnostics
- getAudioLevelBarsFromTimeDomainData

The recorder, preview hook, and dictation test re-import from the
new module. ~120 lines removed from the recorder file with zero
behavior change.
