+++ frontend/src/hooks/audio/__tests__/useAudioDictationRecorder.test.ts
@@ -1,10 +1,8 @@
 import { renderHook, act, waitFor } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import {
-  getAudioLevelBarsFromTimeDomainData,
-  useAudioDictationRecorder,
-} from "../useAudioDictationRecorder";
+import { getAudioLevelBarsFromTimeDomainData } from "../audio-pcm-codec";
+import { useAudioDictationRecorder } from "../useAudioDictationRecorder";
 
 vi.mock("../useAudioInputDevicePreference", () => ({
   useAudioInputDevicePreference: () => ({ selectedAudioInputDeviceId: "" }),
